:lipstick: :fire: Add/Rename files and folders
- Separate constants data into two files as not all data is needed when called upon for specific pages (Team Members and Profile cards)
- Update profile pictures in src folder
- Rename "HomePage" and "ProfilePage" files as "Home" and "OurTeam," respectively, to provide cleaner URL
