--- src/components/ProfileCard.js
@@ -1,86 +1,0 @@
-import Fade from 'react-reveal/Fade';
-
-export default function ProfileCard({
-    id,
-    name,
-    fullBio,
-    imagePath,
-    imageName,
-    projectName,
-    projectDesc,
-    projectImagePath,
-    projectImageName,
-    projectPath,
-    portfolioSite,
-    portfolioName,
-    emailAddress,
-    emailName,
-    LinkedInPage,
-    LinkedInName,
-    otherSite,
-    otherSiteName,
-}) {
-
-    return (
-        <Fade cascade>
-            <div key={id} className="grid grid-cols-1 gap-y-10 gap-x-4 md:px-10">
-                <div className="grid grid-cols-1 md:grid-cols-2 md:gap-x-10">
-                    <img className="img-profiles" src={imagePath} alt={imageName} />
-                    <div className="h-fit">
-                        <h2 className="text-4xl text-left text-darkBlue font-bold border-b-4 w-fit border-lightBlue dark:text-white">{name}</h2>
-                        <p className="para-profiles mt-4">{fullBio}</p>
-                        <ul className="flex gap-x-6 mt-4">
-                            <li>
-                                <a href={portfolioSite} target="_blank" title={portfolioName}>
-                                    <svg className="w-6 h-6 fill-none" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
-                                        <circle cx="12" cy="12" r="12" className="fill-darkBlue dark:fill-lightBlue"/>
-                                        <path d="M18.7765 7.04102H5.16405C4.89706 7.04102 4.68005 7.24927 4.68005 7.50527V8.43923C4.68005 8.48226 4.72872 8.53164 4.76848 8.55161L11.9235 12.4703C11.9421 12.4806 11.9629 12.4853 11.9837 12.4853C12.0049 12.4853 12.0263 12.4801 12.0451 12.4694L18.9838 8.55543C19.0228 8.53455 19.1262 8.47935 19.165 8.45393C19.212 8.42325 19.2605 8.39547 19.2605 8.34046V7.50509C19.2605 7.24927 19.0435 7.04102 18.7765 7.04102Z" className="fill-white"/>
-                                        <path d="M19.2001 9.8474C19.1622 9.82634 19.116 9.82725 19.0779 9.84813L15.1512 12.0634C15.1198 12.0808 15.098 12.1117 15.0927 12.1465C15.088 12.1812 15.0994 12.2166 15.1238 12.2422L19.051 16.3014C19.0743 16.3255 19.1061 16.3384 19.1395 16.3384C19.1544 16.3384 19.1696 16.3357 19.1838 16.3306C19.2302 16.313 19.2605 16.2703 19.2605 16.2224V9.94817C19.2605 9.90641 19.2375 9.8681 19.2001 9.8474Z" className="fill-white"/>
-                                        <path d="M14.0197 12.8074C13.9811 12.7673 13.919 12.7584 13.8699 12.786L12.296 13.674C12.1112 13.7782 11.8745 13.7791 11.6885 13.6778L10.3034 12.9191C10.2576 12.8942 10.1998 12.9002 10.1612 12.9347L4.90126 17.6128C4.87324 17.6379 4.85923 17.674 4.86339 17.7105C4.86756 17.747 4.88971 17.7797 4.92285 17.798C5.00465 17.8441 5.08361 17.8661 5.1639 17.8661H18.6372C18.6854 17.8661 18.729 17.8387 18.7479 17.7962C18.7673 17.7541 18.7585 17.7049 18.7256 17.6711L14.0197 12.8074Z" className="fill-white"/>
-                                        <path d="M9.00381 12.3813C9.03222 12.3564 9.04642 12.3197 9.04225 12.2832C9.03809 12.2466 9.01536 12.2141 8.98204 12.1955L4.86051 9.93837C4.82378 9.91803 4.77738 9.9184 4.73989 9.93909C4.70278 9.95979 4.68005 9.9981 4.68005 10.0393V15.9619C4.68005 16.0078 4.70846 16.0496 4.75239 16.0679C4.76791 16.0746 4.78439 16.0779 4.80105 16.0779C4.83078 16.0779 4.86051 16.0672 4.88323 16.0468L9.00381 12.3813Z" className="fill-white"/>
-                                    </svg>
-                                </a>
-                            </li>
-                            <li>
-                                <a href={emailAddress} target="_blank" title={emailName}>
-                                    <svg className="w-6 h-6 fill-none" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
-                                        <circle cx="12" cy="12" r="12" className="fill-darkBlue dark:fill-lightBlue"/>
-                                        <path d="M18.7765 7.04102H5.16405C4.89706 7.04102 4.68005 7.24927 4.68005 7.50527V8.43923C4.68005 8.48226 4.72872 8.53164 4.76848 8.55161L11.9235 12.4703C11.9421 12.4806 11.9629 12.4853 11.9837 12.4853C12.0049 12.4853 12.0263 12.4801 12.0451 12.4694L18.9838 8.55543C19.0228 8.53455 19.1262 8.47935 19.165 8.45393C19.212 8.42325 19.2605 8.39547 19.2605 8.34046V7.50509C19.2605 7.24927 19.0435 7.04102 18.7765 7.04102Z" className="fill-white"/>
-                                        <path d="M19.2001 9.8474C19.1622 9.82634 19.116 9.82725 19.0779 9.84813L15.1512 12.0634C15.1198 12.0808 15.098 12.1117 15.0927 12.1465C15.088 12.1812 15.0994 12.2166 15.1238 12.2422L19.051 16.3014C19.0743 16.3255 19.1061 16.3384 19.1395 16.3384C19.1544 16.3384 19.1696 16.3357 19.1838 16.3306C19.2302 16.313 19.2605 16.2703 19.2605 16.2224V9.94817C19.2605 9.90641 19.2375 9.8681 19.2001 9.8474Z" className="fill-white"/>
-                                        <path d="M14.0197 12.8074C13.9811 12.7673 13.919 12.7584 13.8699 12.786L12.296 13.674C12.1112 13.7782 11.8745 13.7791 11.6885 13.6778L10.3034 12.9191C10.2576 12.8942 10.1998 12.9002 10.1612 12.9347L4.90126 17.6128C4.87324 17.6379 4.85923 17.674 4.86339 17.7105C4.86756 17.747 4.88971 17.7797 4.92285 17.798C5.00465 17.8441 5.08361 17.8661 5.1639 17.8661H18.6372C18.6854 17.8661 18.729 17.8387 18.7479 17.7962C18.7673 17.7541 18.7585 17.7049 18.7256 17.6711L14.0197 12.8074Z" className="fill-white"/>
-                                        <path d="M9.00381 12.3813C9.03222 12.3564 9.04642 12.3197 9.04225 12.2832C9.03809 12.2466 9.01536 12.2141 8.98204 12.1955L4.86051 9.93837C4.82378 9.91803 4.77738 9.9184 4.73989 9.93909C4.70278 9.95979 4.68005 9.9981 4.68005 10.0393V15.9619C4.68005 16.0078 4.70846 16.0496 4.75239 16.0679C4.76791 16.0746 4.78439 16.0779 4.80105 16.0779C4.83078 16.0779 4.86051 16.0672 4.88323 16.0468L9.00381 12.3813Z" className="fill-white"/>
-                                    </svg>
-                                </a>
-                            </li>
-                            <li>
-                                <a href={LinkedInPage} target="_blank" title={LinkedInName}>
-                                    <svg className='w-6 h-6 fill-none' viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
-                                        <circle cx="12" cy="12" r="12" className="fill-darkBlue dark:fill-lightBlue"/>
-                                        <path d="M18.1805 5H5.75637C5.16222 5 4.68005 5.45145 4.68005 6.00736V17.9716C4.68005 18.5283 5.16222 18.9803 5.75637 18.9803H18.1805C18.7751 18.9803 19.2605 18.5283 19.2605 17.9716V6.00736C19.2605 5.45145 18.7751 5 18.1805 5ZM9.00436 16.9131H6.84079V10.2416H9.00436V16.9131ZM7.92258 9.32942C7.2294 9.32942 6.66907 8.79079 6.66907 8.12673C6.66907 7.46364 7.2294 6.92501 7.92258 6.92501C8.61414 6.92501 9.17589 7.46364 9.17589 8.12673C9.17589 8.79079 8.61414 9.32942 7.92258 9.32942ZM17.1038 16.9131H14.9437V13.6689C14.9437 12.895 14.9279 11.8999 13.8196 11.8999C12.694 11.8999 12.5225 12.7427 12.5225 13.6128V16.9133H10.3607V10.2416H12.4354V11.1527H12.465C12.7538 10.6284 13.4593 10.0752 14.5115 10.0752C16.6998 10.0752 17.104 11.4566 17.104 13.2536V16.9131H17.1038Z" className="fill-white"/>
-                                    </svg>
-                                </a>
-                            </li>
-                            <li>
-                                <a href={otherSite} target="_blank" title={otherSiteName}>
-                                    <svg className="w-6 h-6 fill-none" viewBox="0 0 23 22" xmlns="http://www.w3.org/2000/svg">
-                                        <path d="M11.4785 0.0287163C8.5091 -0.00145781 5.64921 1.14858 3.52735 3.22607C1.40549 5.30356 0.195289 8.13853 0.16272 11.1079C0.174161 13.4546 0.929049 15.7372 2.31896 17.628C3.70886 19.5188 5.66227 20.9204 7.89861 21.6316C8.46439 21.7345 8.67014 21.395 8.67014 21.0967C8.67014 20.7983 8.67014 20.14 8.67014 19.2141C5.52229 19.8828 4.85363 17.7328 4.85363 17.7328C4.64408 17.0578 4.19858 16.4809 3.5986 16.1074C2.5699 15.4285 3.6809 15.4388 3.6809 15.4388C4.03648 15.4863 4.37667 15.6137 4.67598 15.8115C4.97529 16.0092 5.22594 16.2722 5.40913 16.5806C5.72837 17.1342 6.25219 17.5403 6.86785 17.7115C7.48351 17.8828 8.14178 17.8054 8.701 17.4962C8.75851 16.9328 9.01361 16.4081 9.42109 16.0148C6.91104 15.7371 4.27755 14.7907 4.27755 10.5421C4.25511 9.43387 4.6646 8.36034 5.41942 7.54857C5.07565 6.59709 5.11624 5.54909 5.53258 4.62704C5.53258 4.62704 6.48927 4.32872 8.6187 5.75862C10.4723 5.26468 12.423 5.26468 14.2766 5.75862C16.4369 4.32872 17.3627 4.62704 17.3627 4.62704C17.7791 5.54909 17.8196 6.59709 17.4759 7.54857C18.2482 8.34546 18.6799 9.41178 18.6795 10.5215C18.6795 14.7804 16.0254 15.7165 13.5359 15.9943C13.8085 16.2589 14.0191 16.5807 14.1525 16.9364C14.2859 17.2922 14.3388 17.673 14.3075 18.0517V21.0864C14.3075 21.0864 14.5132 21.7345 15.079 21.6213C17.309 20.9057 19.2558 19.5038 20.6414 17.6157C22.027 15.7275 22.7805 13.4498 22.7943 11.1079C22.7617 8.13853 21.5515 5.30356 19.4297 3.22607C17.3078 1.14858 14.4479 -0.00145781 11.4785 0.0287163Z" className="fill-darkBlue dark:fill-lightBlue"/>
-                                    </svg>
-                                </a>
-                            </li>
-                        </ul>
-                    </div> 
-                </div>
-                <div className="grid grid-cols-1 md:grid-cols-2 md:gap-x-10">
-                    <img className="img-profiles" src={projectImagePath} alt={projectImageName} />
-                    <div>
-                        <h3 className="text-darkBlue text-sm lg:text-2xl font-base leading-5 dark:text-white">{projectName}</h3>
-                        <p className="para-profiles mt-2">{projectDesc}</p>
-                        <a href={projectPath} target="_blank" title={projectName}>
-                            <button className='btn-profiles' title="View Project">View Work</button>
-                        </a>
-                    </div>
-                </div>
-            </div>
-        </Fade>
-    );
-}
