+++ src/data/Expertise.js
@@ -1,0 +1,36 @@
+const dataExpertise = [
+	{
+		id: 1,
+		title: "Brand Identity",
+		paragraph:
+			"Be recognized instantly by anyone and everyone. Let your brand speak for itself with our support and tools to stand apart from your competitors.",
+		imagePath: "assets/images/icons/brand_identity.png",
+		imageName: "Web browser with light bulb",
+	},
+	{
+		id: 2,
+		title: "Web Development",
+		paragraph:
+			"Build your digital presence with a custom website design. With our help, you can deliver a captivating experience for every visitor to your site to keep them coming back.",
+		imagePath: "assets/images/icons/website_development.png",
+		imageName: "Web browser with closing code tag icon",
+	},
+	{
+		id: 3,
+		title: "Digital Design",
+		paragraph:
+			"From mobile to tablet to desktop screens, we help to make designs for everyone to experience. Using animations, videos, and graphic designs, we bring your vision to life.",
+		imagePath: "assets/images/icons/digital_design.png",
+		imageName: "Design tools icons",
+	},
+	{
+		id: 4,
+		title: "Marketing Analysis",
+		paragraph:
+			"Get noticed by being in the top-ranked searches when a potential customer searches for you. Our SEO and social media marketing strategies will help you rank higher and stand out effortlessly.",
+		imagePath: "assets/images/icons/marketing_analytics.png",
+		imageName: "Graph Analytics icon",
+	},
+];
+
+export { dataExpertise };
